Better hash

+++ src/vivify.cpp
@@ -397,10 +397,10 @@ void Internal::flush_vivification_schedule (Vivifier &vivifier) {
 // we schedule a clause to be vivified.  For redundant clauses we only try
 // to vivify them if they are likely to survive the next 'reduce' operation.
 
-uint64_t clause_hash(const Clause* c) {
+size_t clause_hash(const Clause* c) {
   size_t h = 17;
   for (const auto &lit : *c)
-    hash_combine(h, (size_t)lit);
+    hash_combine(h, lit);
   return h;
 }
 
@@ -413,8 +413,8 @@ bool Internal::consider_to_vivify_clause (Clause *c, bool redundant_mode) {
     return false;
   if (opts.vivifyonce >= 2 && !c->redundant && c->vivified)
     return false;
-  if (opts.vivifyonly && (int) (clause_hash(c) % (uint64_t) opts.vivifyonlycount) == opts.vivifyonlyid)
-    return true; // erly exit, we don't care for likely clauses, no reduce is performed
+  if (opts.vivifyonly && clause_hash(c) % opts.vivifyonlycount == static_cast<size_t>(opts.vivifyonlyid))
+    return true; // early exit, we don't care for likely clauses, no reduce is performed
   if (c->redundant && !likely_to_be_kept_clause (c))
     return false;
   return true;
